bugfix at dpmCTRL of Voltage data format; add ctrl

+++ lib/dpm86CTRL/src/dpm86CTRL.h
@@ -17,8 +17,8 @@ class dpm86CTRL {
       void setup(int EnPin);                                    // only enable pin   
       void handle();
       void SetResponse();                                       // read response from Bus
-      void setVoltage(int voltage);                             // set voltage output of DPM86xx
-      void setCurrent(int current);                             // set current output of DPM86xx
+      void setVoltage(float voltage);                             // set voltage output of DPM86xx
+      void setCurrent(float current);                             // set current output of DPM86xx
       float readVoltage();
       float readCurrent();
       float readTemp();
@@ -31,7 +31,7 @@ class dpm86CTRL {
     int _RxPin;
     int _TxPin;
     int _EnPin;
-    int _voltage, _current;
+    float _voltage, _current;
     String  _cmd, _ad, _set, _value, _v, _c, _sendOut;          
     String _incomeStr;
     boolean _response;
